# feat: setup new log system and record logs from LLD internal thread