--- a/apps/ledger-live-desktop/src/internal/live-common-setup.ts
+++ b/apps/ledger-live-desktop/src/internal/live-common-setup.ts
@@ -7,6 +7,15 @@ import { retry } from "@ledgerhq/live-common/promise";
 import TransportNodeHidSingleton from "@ledgerhq/hw-transport-node-hid-singleton";
 import TransportHttp from "@ledgerhq/hw-transport-http";
 import { DisconnectedDevice } from "@ledgerhq/errors";
+import { TraceContext, listen as listenLogs } from "@ledgerhq/logs";
+import { ForwardToMainLogger } from "./logger";
+
+const forwardToMainLogger = ForwardToMainLogger.getLogger();
+
+// Listens to logs from the internal threads, and forwards them to the main thread
+listenLogs(log => {
+  forwardToMainLogger.log(log);
+});
 
 /* eslint-disable guard-for-in */
 for (const k in process.env) {
@@ -31,8 +40,8 @@ if (getEnv("DEVICE_PROXY_URL")) {
 } else {
   registerTransportModule({
     id: "hid",
-    open: () =>
-      retry(() => TransportNodeHidSingleton.open(), {
+    open: (id: string, timeoutMs?: number, context?: TraceContext) =>
+      retry(() => TransportNodeHidSingleton.open(id, timeoutMs, context), {
         maxRetry: 4,
       }),
     disconnect: () => Promise.resolve(),