--- a/apps/ledger-live-desktop/src/main/logger.ts
+++ b/apps/ledger-live-desktop/src/main/logger.ts
@@ -119,14 +119,27 @@ export class ConsoleLogger {
       return;
     }
 
-    if (context) {
-      // Displays the tracing context before the message for better readability in the console
-      console.log(
-        `------------------------------\n${type}: ${JSON.stringify(context)}\n${message || ""}\n`,
-        rest,
-      );
-    } else {
-      console.log(`------------------------------\n${type}: ${message || ""}`, rest);
+    try {
+      if (context) {
+        // Displays the tracing context before the message for better readability in the console
+        // eslint-disable-next-line no-console
+        console.log(
+          `------------------------------\n${type}: ${JSON.stringify(context, null, 2)}\n${
+            message || ""
+          }\n${JSON.stringify(rest, null, 2)}`,
+        );
+      } else {
+        // eslint-disable-next-line no-console
+        console.log(
+          `------------------------------\n${type}: ${message || ""}${JSON.stringify(
+            rest,
+            null,
+            2,
+          )}`,
+        );
+      }
+    } catch (_e) {
+      console.error("Badly formatted log");
     }
   }
 }