--- a/apps/ledger-live-desktop/src/renderer/init.tsx
+++ b/apps/ledger-live-desktop/src/renderer/init.tsx
@@ -46,9 +46,31 @@ import { expectOperatingSystemSupportStatus } from "~/support/os";
 import { addDevice, removeDevice, resetDevices } from "~/renderer/actions/devices";
 import { Device } from "@ledgerhq/live-common/hw/actions/types";
 import { listCachedCurrencyIds } from "./bridge/cache";
-if (process.env.VERBOSE) {
-  enableDebugLogger();
+import { LogEntry } from "winston";
+
+const { VERBOSE } = process.env;
+/**
+ * Sets up a debug console printing of logs (from the renderer thread)
+ *
+ * Usage: a filtering (only on console printing) on Ledger libs are possible:
+ * - VERBOSE="apdu,hw,transport,hid-verbose" : filtering on a list of log `type` separated by a `,`
+ * - VERBOSE=1 or VERBOSE=true : to print all logs
+ */
+if (VERBOSE) {
+  const everyLogs = VERBOSE === "true" || VERBOSE === "1";
+  const filters = everyLogs ? [] : VERBOSE.split(",");
+
+  // eslint-disable-next-line no-console
+  console.log(
+    `Logs console display setup (renderer thread): ${JSON.stringify({
+      everyLogs,
+      filters,
+      verbose: VERBOSE,
+    })}`,
+  );
+  enableDebugLogger((log: LogEntry) => everyLogs || (log?.type && filters.includes(log.type)));
 }
+
 const rootNode = document.getElementById("react-root");
 const TAB_KEY = 9;
 async function init() {