… .changeset/lazy-months-smash.md
… .changeset/little-squids-repair.md
… .changeset/purple-chefs-hear.md
… .changeset/thirty-coats-dance.md
… apps/ledger-live-mobile/src/live-common-setup.ts
… apps/ledger-live-mobile/src/logger.ts
… libs/ledger-live-common/src/deviceSDK/tasks/toggleOnboardingEarlyCheck.ts
… libs/ledger-live-common/src/deviceSDK/transports/core.ts
… libs/ledger-live-common/src/hw/deviceAccess.ts
… libs/ledger-live-common/src/hw/index.ts
… libs/ledgerjs/packages/hw-transport-node-hid-noevents/README.md
… libs/ledgerjs/packages/hw-transport-node-hid-noevents/src/TransportNodeHid.ts
… libs/ledgerjs/packages/hw-transport-node-hid-singleton/README.md
… libs/ledgerjs/packages/hw-transport-node-hid-singleton/src/TransportNodeHid.ts
… libs/ledgerjs/packages/hw-transport-node-hid-singleton/src/listenDevices.ts
… libs/ledgerjs/packages/hw-transport/README.md
… libs/ledgerjs/packages/hw-transport/package.json
… libs/ledgerjs/packages/hw-transport/src/Transport.ts
… libs/ledgerjs/packages/logs/README.md
… libs/ledgerjs/packages/logs/src/index.ts
… libs/ledgerjs/packages/react-native-hw-transport-ble/src/BleTransport.ts
… libs/ledgerjs/packages/react-native-hw-transport-ble/src/monitorCharacteristic.ts
… pnpm-lock.yaml