<!--
Thank you for your contribution! 👍
Please make sure to read CONTRIBUTING.md if you have not already.
Disclaimer: Pull Requests that do not comply with the rules will be arbitrarily closed.
-->

### 📝 Description

#### `ledgerjs/logs`
- new `trace` and `LocalTracer` definitions, **without breaking the behavior** of the existing `log` function

#### `hw-transport`, `hw-transport-node-hid-noevents`, `hw-transport-node-hid-singleton`, `react-native-hw-transport-ble`, `ledger-live-common` 

- Usage of new tracing system in some transports
- The tracing helps keeping a context (for ex a `job id`) that is propagated to other logs, creating a (simple) tracing span

Note on the "tracing": for now added a simple "context" that is propagated from `withDevice` to called functions, mainly from Transport. This is a preview of what could be done to have a real tracing system in the DeviceSDK

Note: the optional `openTimeoutMs` on the different `open` function is necessary for legacy (used by one of the Transport), and will actually be used in another PR. Don't mind it during the code review. But it should only by passed to a couple of other functions, not directly used (in this PR).

#### LLD:
- Logs from the `internal` thread are forwarded to the `main` thread
- The `main` thread records them, and can export them
  - on export, recorded logs from the `renderer` thread are sent to the `main` thread, and the `main` thread expand (not a merge, just pushed at the end) them with the logs it recorded from the `internal` thread
- If `VERBOSE` env var is set, filtered logs can be stdout from the main thread
  - Same from the `renderer` thread
- Setup simple tracing system (context) on LLD

Usage with `VERBOSE`:
```
VERBOSE="apdu,hw,hid-verbose,usb-detection" pnpm dev:lld
```

#### LLM:
- Setup simple tracing system on LLM with context
- If `VERBOSE` env var is set, filtered logs can be stdout from the main thread

Usage with `VERBOSE`: in a `.env` file, add (for ex)
```
VERBOSE="apdu,hw,ble-verbose,ble-error,transport"
```
You will need to re-build fully the app so the new `.env` file is transfered.

### ❓ Context

- **Impacted projects**: `hw-transport`, `hw-transport-node-hid-noevents`, `hw-transport-node-hid-singleton`, `react-native-hw-transport-ble`, `ledger-live-common` , `ledgerjs/logs`, LLM and LLD
- **Linked resource(s)**:  [LIVE-9405](https://ledgerhq.atlassian.net/browse/LIVE-9405)

### ✅ Checklist

- [ ] **Test coverage** <!-- Are your changes covered by tests? Features must be tested, bugfixes must include a test that would have detected the issue. -->
- [x] **Atomic delivery** <!-- Is this pull request standalone? In order words, does it depend on nothing else? Please explain if not checked. -->
- [ ] **No breaking changes** <!-- If there are breaking changes, please explain why. -->

### 📸 Demo

#### LLD performance

Non-scientific manual measurements (on dev build): beginning of the fw update until the end of the fw transfer:

<details>
  <summary>Measurements fw update transfer</summary>

With new log system (without `VERBOSE`):
- 47.909 s
- 48.228 s
- 46.515 s
- 47.181 s
- 47.807 s

Without new log system:
- 47.311 s
- 48.783 s
- 48.184 s
- 47.092 s
- 47.667 s

-> no obvious performance decline
</details>

#### Exported logs from LLD

<details>
  <summary>Example separation `renderer` and `internal` thread</summary>

<img width="885" alt="LLD_export_logs_separation" src="https://github.com/LedgerHQ/ledger-live/assets/15096913/9cf189c6-ec9f-4c28-bb68-b8c40fc28a67">


</details>

#### Stdout debug logs on LLD with `VERBOSE`

<details>
  <summary>LLD logs on terminal</summary>

<img width="722" alt="LLD_stdout_logs_VERBOSE" src="https://github.com/LedgerHQ/ledger-live/assets/15096913/0db4577b-1799-476d-9947-c0749625f310">


</details>

#### Stdout debug logs on LLM with `VERBOSE`

<details>
  <summary>LLM logs on terminal</summary>

<img width="431" alt="LLM_stdout_ex" src="https://github.com/LedgerHQ/ledger-live/assets/15096913/2e3faef2-d405-4364-b5ed-7aaf8c794315">

</details>

### 🚀 Expectations to reach



[LIVE-9405]: https://ledgerhq.atlassian.net/browse/LIVE-9405?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ